refactor(gen_keys): gen_keys from oauth combined into auth gen_keys
