+++ scripts/start-local.sh
@@ -3,7 +3,6 @@
 set -euo pipefail
 
 node ./scripts/gen_keys.js
-node ./fxa-oauth-server/scripts/gen_keys.js
 node ./scripts/gen_vapid_keys.js
 node ./test/mail_helper.js &
 MH=$!
